+++ CHANGELOG.md
@@ -1,3 +1,11 @@
 ﻿# Changelog
-## 1.0.0 (December 22nd, 2017)
+## 1.0.1 (December 23rd, 2017)
+- Moved all classes into their proper namespace, `Aeonix`
+- Renamed 2 classes to prevent namespacing issues:
+  - `Command` > `CommandBase`
+  - `CommandHandler` > `CommandHandlerBase`
+- Moved `CommandHandlerBase` constructor(s) into a method called `Init`
+- [Code Cleanup](https://github.com/TakeTenGaming/Aeonix/issues/1)
+
+### 1.0.0 (December 22nd, 2017)
 - Initial Framework release